pyproject.toml
[project]
name = "llmgine"
version = "0.0.1"
description = "An application framework for building LLM-powered applications."
authors = [{ name = "Nathan Luo", email = "nathanluo13@gmail.com" }]
readme = "README.md"
keywords = ['python']
requires-python = ">=3.11,<4.0"
classifiers = [
    "Intended Audience :: Developers",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "aiofiles>=24.1.0",
    "aiosqlite>=0.21.0",
    "discord>=2.3.2",
    "dotenv>=0.9.9",
    "icecream>=2.1.4",
    "instructor>=1.7.9",
    "ipdb>=0.13.13",
    "ipython>=8.18.1",
    "litellm>=1.63.12",
    "load-dotenv>=0.1.0",
    "mcp>=1.10.1",
    "notion-client>=2.3.0",
    "pandas>=2.2.3",
    "prompt-toolkit>=3.0.51",
    "psutil>=7.0.0",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.11.3",
    "pyreadline>=2.1",
    "pytest>=8.3.5",
    "pytest-asyncio>=0.25.3",
    "python-dotenv>=1.1.0",
    "rich>=13.9.4",
    "sqlalchemy>=2.0.40",
    "textual>=2.1.2",
]

[project.urls]
Homepage = "https://nathan-luo.github.io/llmgine/"
Repository = "https://github.com/nathan-luo/llmgine"
Documentation = "https://nathan-luo.github.io/llmgine/"

[project.optional-dependencies]
opentelemetry = [
    "opentelemetry-api>=1.20.0",
    "opentelemetry-sdk>=1.20.0",
    "opentelemetry-exporter-otlp>=1.20.0",
    "opentelemetry-semantic-conventions>=0.41b0"
]

[dependency-groups]
dev = [
    "pytest>=7.2.0",
    "pytest-asyncio>=0.23.0",
    "pre-commit>=2.20.0",
    "tox-uv>=1.11.3",
    "deptry>=0.22.0",
    "mypy>=0.991",

    "ruff>=0.9.2",
    "mkdocs>=1.4.2",
    "mkdocs-material>=8.5.10",
    "mkdocstrings[python]>=0.26.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/llmgine", "tools"]


[tool.mypy]
files = ["src"]
disallow_untyped_defs = true
disallow_any_unimported = true
no_implicit_optional = true
check_untyped_defs = true
warn_return_any = true
warn_unused_ignores = true
show_error_codes = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-sv --log-cli-level=0"

[tool.ruff]
target-version = "py39"
line-length = 90
fix = true

[tool.ruff.lint]
select = [
    # flake8-2020
    "YTT",
    # flake8-bandit
    "S",
    # flake8-bugbear
    "B",
    # flake8-builtins
    "A",
    # flake8-comprehensions
    "C4",
    # flake8-debugger
    "T10",
    # flake8-simplify
    "SIM",
    # isort
    "I",
    # mccabe
    "C90",
    # pycodestyle
    "E",
    "W",
    # pyflakes
    "F",
    # pygrep-hooks
    "PGH",
    # ruff
    "RUF",
]
ignore = [
    # LineTooLong
    "E501",
    # DoNotAssignLambda
    "E731",
    # Blank line contains whitespace
    "W293",
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101"]

[tool.ruff.format]
preview = true
